Move local illuminance cluster and custom PowerConfiguration class to main Xiaomi file
They'll also be used by other devices (in future PRs)
